goalpilot/interview-predictor/requirements.txt
# Core API
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ML/NLP
torch==2.1.1
transformers==4.35.2
sentencepiece==0.2.0
protobuf==4.25.1

# Audio processing
openai-whisper==20231117
soundfile==0.12.1

# AI Feedback
google-generativeai==0.3.1

# Utilities
numpy==1.26.0
